pyproject.toml
[project]
name = "definet"
version = "0.1.3"
description = "a tool that will give you the definition of any word you like."
authors = [{ name = "sigmacodeslol", email = "tcrrxzz@gmail.com" }]
license = "MIT"
readme = "README.rst"
requires-python = ">=3.12"
dependencies = ["pyenchant", "requests"]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.scripts]
definet = "definet.__main__:main"

[project.urls]
Repository = "https://github.com/sigmacodeslol/definet.git"


[build-system]
requires = ["setuptools >= 77.0.3", "wheel"]
build-backend = "setuptools.build_meta"
